requirements.txt
pywebview>=4.0
requests>=2.28
pyobjc-framework-Cocoa>=9.0
